Rename omitBoxProps.js to boxProps.js
