+++ components/box/index.js
@@ -1,5 +1,5 @@
 import Box from './Box';
-import boxProps, { omitBoxProps, pickBoxProps } from './omitBoxProps';
+import boxProps, { omitBoxProps, pickBoxProps } from './boxProps';
 
 export default Box;
 export { Box, boxProps, omitBoxProps, pickBoxProps };
